Bug Fix for PendingValidation
When setting the state to PendingValidation, then the guide UI was not dismissed once the Rule had been satisfied.
