+++ GuidedEditText/build.gradle
@@ -10,8 +10,8 @@ android {
     defaultConfig {
         minSdk 24
         targetSdk 30
-        versionCode 2
-        versionName "1.6.0"
+        versionCode 3
+        versionName "1.6.1"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
