+++ app/src/main/res/layout/activity_main.xml
@@ -167,6 +167,7 @@
         android:paddingEnd="16dp"
         android:paddingTop="8dp"
         android:paddingBottom="8dp"
+        android:inputType="textPassword"
         android:textSize="16sp"
         android:layout_marginTop="16dp"
         android:background="@drawable/rounded_corner"
